fixing warnings and errors

+++ src/core/CrystalController.cpp
@@ -20,7 +20,7 @@ void CrystalController::onMessage(uint16_t message)
             // TODO, client should be able to create entities locally and have everything line up correctly on the registry side. Only way to do that though... is to sync _nextNetworkEntityID
             if (_entity->isServer())
             {
-                NW_SRVR->registerNewEntity('EXPL', _entity->position()._x, _entity->position()._y);
+                NW_SRVR->registerNewEntity('EXPL', (uint32_t)_entity->position()._x, (uint32_t)_entity->position()._y);
                 NW_SRVR->registerNewEntity('MON1', rand() % 24 + 6, rand() % 32 + 12);
                 _entity->requestDeletion();
             }
